Refactoring: code separation
