+++ App.js
@@ -1,11 +1,9 @@
 import "react-native-gesture-handler";
 import React from "react";
-import { StyleSheet } from "react-native";
 import Routes from "./Routes";
 import { Provider } from "react-redux";
 import { createStore } from "redux";
 import reducers from "./store/reducers";
-import * as Localization from "expo-localization";
 
 const store = createStore(reducers);
 
@@ -16,13 +14,3 @@ export default function App() {
     </Provider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    width: "100%",
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
