+++ RootNavigation.js
@@ -4,12 +4,10 @@ export const isReadyRef = React.createRef();
 
 export const navigationRef = React.createRef();
 
-export function navigate(name, params) {
-  if (navigationRef.current) {
-    setTimeout(() => {
-      navigationRef.current.navigate(name, params);
-    }, 0);
+export const navigate = (name, params) => {
+  if (navigationRef.current && isReadyRef.current) {
+    navigationRef.current.navigate(name, params);
   } else {
-    console.log(`navigation isn't ready yet`);
+    console.log("navigation isn't ready yet");
   }
 }
